pyproject.toml
[project]
name = "voice-browser"
version = "0.1.0"
description = "Voice-controlled web browser using Fish Audio and browser-use"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.9.7",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "python-multipart>=0.0.9",
    "faster-whisper>=1.0.3",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "sounddevice>=0.5.0",
    "pydantic>=2.9.0",
    "pydub>=0.25.1",
    "playwright>=1.48.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
